refactor: adapt code to p0tion files
