+++ .gitignore
@@ -39,3 +39,5 @@ p0tion-artifacts
 p0tion-repo-artifacts
 
 contract-verification-keys-hex.txt
+
+snark-artifacts-repo-files
