+++ package.json
@@ -16,16 +16,15 @@
         "create:files": "ts-node ./scripts/createFiles.ts",
         "remove:files": "ts-node ./scripts/removeCreatedFiles.ts",
         "copy:files": "ts-node ./scripts/copyArtifactsFiles.ts",
-        "remove:artifacts": "ts-node ./scripts/removeArtifactsFolder.ts",
         "merge:verification-keys": "ts-node ./scripts/mergeVerificationKeys.ts",
-        "get:wasm-zkey-files": "ts-node ./scripts/getWasmZkeyFiles.ts",
         "get:ceremony-files": "ts-node ./scripts/getCeremonyFiles.ts",
         "get:contract-keys": "ts-node ./scripts/getContractVerifierKeys.ts",
         "export-verification-key": "chmod u+x ./scripts/exportVerificationKey.sh && ./scripts/exportVerificationKey.sh",
-        "rename-artifacts": "ts-node ./scripts/renameArtifacts.ts 1 32",
+        "p0tion-rename-artifacts": "ts-node ./scripts/p0tion-rename-artifacts.ts 1 32",
         "convert-verifications-keys-to-hex": "ts-node ./scripts/convertVKPtsBigIntToHex.ts",
         "export-keys": "yarn export-verification-key && yarn merge:verification-keys 1 32 && yarn convert-verifications-keys-to-hex",
-        "run:all": "yarn remove:files 1 32 && yarn create:files 1 32 && yarn execute 1 32 && yarn remove:artifacts && yarn copy:files 1 32 && yarn merge:verification-keys 1 32 && yarn get:wasm-zkey-files 1 32 && yarn get:ceremony-files 1 32 && yarn get:contract-keys"
+        "p0tion-get-snark-artifacts-repo-files": "ts-node ./scripts/generate-snark-artifacts-repo-files.ts 1 32 && yarn export-keys",
+        "run:all": "yarn remove:files 1 32 && yarn create:files 1 32 && yarn execute 1 32 && yarn copy:files 1 32 && yarn merge:verification-keys 1 32 && yarn get:ceremony-files 1 32 && yarn get:contract-keys"
     },
     "devDependencies": {
         "@types/chai": "^4.3.11",
