+++ scripts/p0tion-rename-artifacts.ts
