--- scripts/removeArtifactsFolder.ts
@@ -1,3 +1,0 @@
-import { rmSync } from "fs"
-
-rmSync("zk-artifacts", { recursive: true, force: true })
